Extracted directions map logic from Index.tsx into useNavigationCamera hook
